Time: 1 ms (99.8%), Space: 42.7 MB (63.07%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.java
@@ -1,7 +1,7 @@
 class Solution {
     public int reverse(int x) {
         if(x>=0 && x<=9) return x;
-        int rev=0;
+        long rev=0;
         while(x!=0){
             int last=x%10;
             rev=rev*10+last;
@@ -10,6 +10,6 @@ class Solution {
                 return 0;
             }
         }
-        return rev;
+        return (int)rev;
     }
 }
